+++ CHANGELOG.md
@@ -15,6 +15,7 @@ for further information about branching and tagging conventions.
 - Sign in page
 - Lodash
 - `AppButton` component
+- `StatusMessage` component
 
 ## [0.1.0-rc.1] - 2021-05-28
 #### Initial project release